CounteR/requirements.txt
amqp==5.0.9
analytics-python==1.4.0
aniso8601==9.0.1
attrs==19.3.0
Automat==0.8.0
backoff==1.10.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
botocore==1.23.43
cached-property==1.5.1
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.6
constantly==15.1.0
cryptography==38.0.4
cycler==0.11.0

Deprecated==1.2.13

docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
ephem==3.7.7.1
ffmpy==0.3.0
Flask==2.0.2
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-restx==0.5.1
fonttools==4.28.2
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gradio==2.4.6
httplib2==0.20.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
instagram==1.3.4
instaloader==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
json-stream==1.3.0
json2html==1.3.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2


launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
markdown2==2.4.1
MarkupSafe==2.0.1
matplotlib==3.5.0
monotonic==1.6
more-itertools==4.2.0
numpy==1.21.4
oauthlib==3.1.0
packaging==21.3
pandas==1.3.4
paramiko==2.8.1
pendulum==2.1.2
pexpect==4.6.0
Pillow==8.4.0
pip-chill==1.0.1
prompt-toolkit==3.0.24
protobuf==3.19.1
pure-sasl==0.6.2
py4j==0.10.9.2
pyaes==1.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.11.0
pydub==0.25.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.6
pyrsistent==0.15.5
pyserial==3.4
python-dateutil==2.8.2
python-memcached==1.58
pytube==11.0.2
# git+https://github.com/sadeghastaneh/pytube.git@patch-1

pytwitter==0.5
pytz==2021.3
pytzdata==2020.1
pywebhdfs==0.4.1
PyYAML==5.3.1
redis==4.1.1
#requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.std.zipfile==0.1.0
s3transfer==0.5.0
sentry-sdk
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
telethon==1.27.0
texttable==1.6.2
thrift==0.15.0
thrift-sasl==0.4.3
toml==0.10.2
tomli==1.2.2
tweepy==4.4.0
uritemplate==4.1.1
#urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
webhdfspy==0.3.5
#websocket-client==0.53.0
Werkzeug<=2.1.2
wrapt==1.13.3
youtube-data-api==0.0.21
youtube_transcript_api==0.5.0
zipp==1.0.0
semantic-version==2.10.0

names==0.3.0
streamlink==6.1.0
emoji==2.2.0

beautifulsoup4==4.12.2
fake-useragent==1.4.0
